+++ imm-20200930/ChangeLog.md
@@ -1,3 +1,8 @@
+2024-07-30 Version: 4.5.4
+- Update API BatchGetFileMeta: add param WithFields.
+- Update API GetFileMeta: add param WithFields.
+
+
 2024-07-25 Version: 4.5.3
 - Generated python 2020-09-30 for imm.